move training configuration for pytorch models into the same file
