+++ external/fv3fit/fv3fit/pytorch/graph/__init__.py
@@ -1,3 +1,3 @@
 from .network import GraphNetwork, GraphNetworkConfig, CubedSphereGraphOperation
-from .train import train_graph_model, TrainingLoopConfig, GraphHyperparameters
+from .train import train_graph_model, AutoregressiveTrainingConfig, GraphHyperparameters
 from .graph_builder import build_dgl_graph, build_graph
